requirements.txt
awswrangler==3.2.1
boto3==1.26.165
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.5.0/fr_core_news_sm-3.5.0-py3-none-any.whl
matplotlib==3.7.1
nltk==3.8.1
pandas==2.0.3
plotly==5.15.0
pyarrow==12.0.1
python-dotenv==1.0.0
scipy==1.11.1
spacy==3.5.4
streamlit==1.24.0
wordcloud==1.9.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
transformers==4.27.1
tweepy==4.14.0
openpyxl==3.1.2
sentencepiece==0.1.99